% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks_show_tables.R
\name{databricks_show_tables}
\alias{databricks_show_tables}
\title{Show tables in a Databricks schema}
\usage{
databricks_show_tables(sc, schema)
}
\arguments{
\item{sc}{A sparklyr connection.}

\item{schema}{A Databricks schema.}
}
\value{
Tables in the schema.
}
\description{
Show tables in a Databricks schema
}
